Adding Adam(W) optimizer
